design-system-extractor/scripts/requirements.txt
playwright==1.56.0
Pillow==12.0.0
numpy==2.3.5
scikit-learn==1.7.2
PyYAML==6.0.3
